add cards & svg to home, add card, add open on sample page
