+++ components/pageSections/HomeContent.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Image from "next/image";
 import Link from "next/link";
+import CvCard from "../resumeComponents/CvCard";
 
 const HomeContent = () => {
     return (
@@ -43,30 +44,54 @@ const HomeContent = () => {
 
             </div>
 
-            <div className='bg-violet-700 flex justify-center'
+            <div className='bg-violet-700 flex-row'
                  style={{width: '100%', height: '1693px'}}
             >
+                <div className='bg-violet-50' style={{height: 1}}></div>
                 <div className='bg-white mt-28 flex justify-center' style={{width: '100%', height: '268px'}}>
-                    {/*<h2 style={{border: '1px solid black'}}>TEXT</h2>*/}
+                        {/*<h2 style={{border: '1px solid black'}}>TEXT</h2>*/}
                 </div>
+                <Link href={'/box/personal'}>
+                    <div className='mt-28 grid grid-cols-3 mx-auto' style={{width: '70%', height: '1040px'}}>
+                        <CvCard/>
+                        <CvCard/>
+                        <CvCard/>
+                        <CvCard/>
+                        <CvCard/>
+                        <CvCard/>
+                    </div>
+                </Link>
             </div>
 
-            <div className=" pb-5 pt-2 container mx-auto mb-16 mt-16">
+            <svg xmlns="http://www.w3.org/2000/svg" width="100%"  className='animation-svg animation'>
+                <defs>
+                    <pattern id="blocks" patternUnits="userSpaceOnUse" width="200" height="200">
+                        <rect width="100" height="100" className="topleft black"/>
+                        <rect width="100" height="100" className="topright white" x="100"/>
+                        <rect width="100" height="100" className="bottomleft white" y="100"/>
+                        <rect width="100" height="100" x="100" y="100" className="bottomright black"/>
+                    </pattern>
+                </defs>
+                <rect width="100%" height="100%" fill="url(#blocks)"/>
+            </svg>
+
+            <div className=" pb-5 pt-2 container mx-auto mb-16 mt-16" style={{marginTop: '-465px'}}>
                 <div
-                    style={{width: '890px', height: '325px'}}
+                    style={{width: '850px', height: '325px'}}
                     className="ml-auto mr-auto mt-12 items-center text-center max-w-[887px] lg:max-w-[887px] space-y-3"
                 >
-                    <div className="text-2xl text-gray-900 font-extrabold lg:text-4xl">Only 2% of resumes make it past
+                    <div className="text-2xl text-black font-extrabold lg:text-4xl">Only 2% of resumes make it past
                         the first round. Be in the top 2%
                     </div>
-                    <p className="text-gray-900 text-md lg:text-xl mb-16">
+                    <p className="text-black text-md lg:text-xl mb-16">
                         Use professional field-tested resume templates that
-                        follow the exact ‘resume rules’ employers look for. Easy to use and done within minutes - try
+                        follow the exact ‘resume rules’ employers look for.
+                        Easy to use and done within minutes - try
                         now for free!
                     </p>
                     <Link href="/box/personal">
                         <button
-                            className="bg-green-500 text-white text-lg xl:text-xl font-semibold rounded-md py-2 px-4 xl:py-4 xl:px-7 hover:bg-green-600 ">Create
+                            className="bg-violet-900 text-white text-lg xl:text-xl font-semibold rounded-md py-2 px-4 xl:py-4 xl:px-7 hover:bg-green-600 ">Create
                             My Resume
                         </button>
                     </Link>
