+++ styles/globals.css
@@ -70,3 +70,66 @@
     from {transform: rotate(0deg);}
     to {transform: rotate(359deg);}
 }
+/*end loading style*/
+
+
+/*background squares */
+@keyframes backgroundshift {
+    49.99% {
+        background: rgb(245,243,255);
+    }
+    50% {
+        background: rgb(109,40, 217);
+    }
+    100% {
+        background: rgb(109,40, 217);
+    }
+}
+
+@keyframes scaleshift {
+    0% {
+        transform: scale(.001);
+        rx: 50;
+    }
+    35% {
+        rx: 40;
+    }
+    50% {
+        transform: scale(1);
+        rx: 0;
+    }
+    100% {
+        transform: scale(1);
+        rx: 0;
+    }
+}
+.animation {
+    animation: backgroundshift 7s alternate infinite;
+}
+.animation-svg { min-height: 50vh; }
+
+pattern rect {
+    animation: scaleshift 7s infinite cubic-bezier(.36,.17,.86,.67);
+}
+.black {
+    fill: rgb(109,40, 217);
+    animation-direction: alternate;
+}
+.white {
+    fill: rgb(245,243,255);
+    animation-direction: alternate-reverse;
+}
+.topleft {
+    transform-origin: 50px 50px;
+}
+.bottomleft {
+    transform-origin: 50px 150px;
+}
+.topright {
+    transform-origin: 150px 50px;
+}
+.bottomright {
+    transform-origin: 150px 150px;
+}
+/*end background squares styles*/
+
